webapp/requirements.txt
Flask
Flask-SocketIO
gevent-websocket
loguru
redis